Zmigrowano baze na Google Cloud.

+++ LearningPlanner 1.0.0/Tasks.cs
@@ -20,6 +20,6 @@ namespace LearningPlanner
         public string Opis { get; set; }
         public System.DateTime DataUtworzenia { get; set; }
         public bool CzyZakonczone { get; set; }
-        public int IDUzytkownika { get; set; }
+        public short IDUzytkownika { get; set; }
     }
 }
